requirements-gcp.txt
prefect-gcp
pydata-google-auth
google-auth
google-auth-oauthlib
google-cloud-bigquery
google-cloud-bigquery-storage
pandas-gbq
google-cloud-storage